# src/Spotless.Web/lib/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Spotless API",
    "description": "Laundry and dry cleaning service API",
    "version": "v1"
  },
  "paths": {
...
    "/api/customers": {
      "get": {
        "tags": [
          "Customers"
        ],
        "summary": "Lists all customers (Admin only)",
        "parameters": [
          {
            "name": "nameFilter",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "emailFilter",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "pageNumber",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Spotless.Application.Dtos.Responses.PagedResponse`1[[Spotless.Application.Dtos.Customer.CustomerDto, Spotless.Application, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null]]"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Spotless.Application.Dtos.Responses.PagedResponse`1[[Spotless.Application.Dtos.Customer.CustomerDto, Spotless.Application, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null]]"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Spotless.Application.Dtos.Responses.PagedResponse`1[[Spotless.Application.Dtos.Customer.CustomerDto, Spotless.Application, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null]]"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/customers/register": {
      "post": {
        "tags": [
          "Customers"
        ],
        "summary": "Registers a new customer account",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Spotless.Application.Features.Customers.Commands.RegisterCustomer.RegisterCustomerCommand"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Spotless.Application.Features.Customers.Commands.RegisterCustomer.RegisterCustomerCommand"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Spotless.Application.Features.Customers.Commands.RegisterCustomer.RegisterCustomerCommand"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Spotless.Application.Dtos.Authentication.AuthResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Spotless.Application.Dtos.Authentication.AuthResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Spotless.Application.Dtos.Authentication.AuthResult"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/customers/dashboard": {
      "get": {
        "tags": [
          "Customers"
        ],
        "summary": "Retrieves customer dashboard with orders and wallet info",
        "parameters": [
          {
            "name": "pageNumber",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Spotless.Application.Dtos.Customer.CustomerDashboardDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Spotless.Application.Dtos.Customer.CustomerDashboardDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Spotless.Application.Dtos.Customer.CustomerDashboardDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/customers/me": {
      "get": {
        "tags": [
          "Customers"
        ],
        "summary": "Retrieves authenticated customer's profile",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Spotless.Application.Dtos.Customer.CustomerDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Spotless.Application.Dtos.Customer.CustomerDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Spotless.Application.Dtos.Customer.CustomerDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Customers"
        ],
        "summary": "Updates authenticated customer's profile",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Spotless.Application.Dtos.Customer.CustomerUpdateRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Spotless.Application.Dtos.Customer.CustomerUpdateRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Spotless.Application.Dtos.Customer.CustomerUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/customers/topup": {
      "post": {
        "tags": [
          "Customers"
        ],
        "summary": "Tops up customer's wallet balance",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Spotless.Application.Dtos.Customer.WalletTopUpRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Spotless.Application.Dtos.Customer.WalletTopUpRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Spotless.Application.Dtos.Customer.WalletTopUpRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Microsoft.AspNetCore.Mvc.ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": {}
      },
...
      "Spotless.Application.Dtos.Authentication.AuthResult": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "accessToken": {
            "type": "string",
            "nullable": true
          },
          "accessTokenExpiration": {
            "type": "string",
            "format": "date-time"
          },
          "refreshToken": {
            "type": "string",
            "nullable": true
          },
          "refreshTokenExpiration": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
...
      "Spotless.Application.Dtos.Customer.CustomerDashboardDto": {
        "type": "object",
        "properties": {
          "totalOrders": {
            "type": "integer",
            "format": "int32"
          },
          "upcomingBookedServices": {
            "type": "integer",
            "format": "int32"
          },
          "walletBalance": {
            "type": "number",
            "format": "double"
          },
          "walletCurrency": {
            "type": "string",
            "nullable": true
          },
          "upcomingOrders": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Spotless.Application.Dtos.Order.OrderDto"
            },
            "nullable": true
          },
          "recentTransactions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Spotless.Application.Dtos.Payment.PaymentDto"
            },
            "nullable": true
          },
          "totalReviewsSent": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "Spotless.Application.Dtos.Customer.CustomerDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "phone": {
            "type": "string",
            "nullable": true
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "street": {
            "type": "string",
            "nullable": true
          },
          "city": {
            "type": "string",
            "nullable": true
          },
          "country": {
            "type": "string",
            "nullable": true
          },
          "zipCode": {
            "type": "string",
            "nullable": true
          },
          "walletBalance": {
            "type": "number",
            "format": "double"
          },
          "walletCurrency": {
            "type": "string",
            "nullable": true
          },
          "type": {
            "$ref": "#/components/schemas/Spotless.Domain.Enums.CustomerType"
          }
        },
        "additionalProperties": false
      },
      "Spotless.Application.Dtos.Customer.CustomerUpdateRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          },
          "phone": {
            "type": "string",
            "nullable": true
          },
          "street": {
            "type": "string",
            "nullable": true
          },
          "city": {
            "type": "string",
            "nullable": true
          },
          "country": {
            "type": "string",
            "nullable": true
          },
          "zipCode": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Spotless.Application.Dtos.Customer.WalletTopUpRequest": {
        "type": "object",
        "properties": {
          "amountValue": {
            "type": "number",
            "format": "double"
          },
          "paymentMethod": {
            "$ref": "#/components/schemas/Spotless.Domain.Enums.PaymentMethod"
          }
        },
        "additionalProperties": false
      },
...
      "Spotless.Application.Dtos.Order.OrderDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "customerId": {
            "type": "string",
            "format": "uuid"
          },
          "driverId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "timeSlotId": {
            "type": "string",
            "format": "uuid"
          },
          "scheduledDate": {
            "type": "string",
            "format": "date-time"
          },
          "pickupLatitude": {
            "type": "number",
            "format": "double"
          },
          "pickupLongitude": {
            "type": "number",
            "format": "double"
          },
          "pickupAddress": {
            "type": "string",
            "nullable": true
          },
          "deliveryLatitude": {
            "type": "number",
            "format": "double"
          },
          "deliveryLongitude": {
            "type": "number",
            "format": "double"
          },
          "deliveryAddress": {
            "type": "string",
            "nullable": true
          },
          "totalPrice": {
            "type": "number",
            "format": "double"
          },
          "currency": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "$ref": "#/components/schemas/Spotless.Domain.Enums.OrderStatus"
          },
          "paymentMethod": {
            "$ref": "#/components/schemas/Spotless.Domain.Enums.PaymentMethod"
          },
          "orderDate": {
            "type": "string",
            "format": "date-time"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Spotless.Application.Dtos.Order.OrderItemDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Spotless.Application.Dtos.Order.OrderItemDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "serviceId": {
            "type": "string",
            "format": "uuid"
          },
          "priceAmount": {
            "type": "number",
            "format": "double"
          },
          "priceCurrency": {
            "type": "string",
            "nullable": true
          },
          "quantity": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "Spotless.Application.Dtos.Payment.PaymentDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "orderId": {
            "type": "string",
            "format": "uuid"
          },
          "amount": {
            "type": "number",
            "format": "double"
          },
          "currency": {
            "type": "string",
            "nullable": true
          },
          "paymentDate": {
            "type": "string",
            "format": "date-time"
          },
          "method": {
            "$ref": "#/components/schemas/Spotless.Domain.Enums.PaymentMethod"
          },
          "status": {
            "$ref": "#/components/schemas/Spotless.Domain.Enums.PaymentStatus"
          }
        },
        "additionalProperties": false
      },
...
      "Spotless.Application.Features.Customers.Commands.RegisterCustomer.RegisterCustomerCommand": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "nullable": true
          },
          "password": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "phone": {
            "type": "string",
            "nullable": true
          },
          "type": {
            "$ref": "#/components/schemas/Spotless.Domain.Enums.CustomerType"
          },
          "street": {
            "type": "string",
            "nullable": true
          },
          "city": {
            "type": "string",
            "nullable": true
          },
          "country": {
            "type": "string",
            "nullable": true
          },
          "zipCode": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "Spotless.Domain.Enums.CustomerType": {
        "enum": [
          0,
          1
        ],
        "type": "integer",
        "format": "int32"
      },
...
      "Spotless.Domain.Enums.OrderStatus": {
        "enum": [
          0,
          1,
          2,
          3,
          4,
          5,
          6,
          7,
          8
        ],
        "type": "integer",
        "format": "int32"
      },
      "Spotless.Domain.Enums.PaymentMethod": {
        "enum": [
          "CreditCard",
          "DebitCard",
          "Fawry",
          "PayPal",
          "Wallet",
          "CashOnDelivery",
          "Paymob",
          "Instapay"
        ],
        "type": "string"
      },
      "Spotless.Domain.Enums.PaymentStatus": {
        "enum": [
          0,
          1,
          2,
          3
        ],
        "type": "integer",
        "format": "int32"
      }
    },
    "securitySchemes": {
      "Bearer": {
        "type": "apiKey",
        "description": "JWT Authorization header using the Bearer scheme. Enter 'Bearer' [space] and then your token",
        "name": "Authorization",
        "in": "header"
      }
    }
  },
  "security": [
    {
      "Bearer": []
    }
  ]
}